Backport Ruby 2.4 fixes
Ruby 2.4 support is included across versions 3.3 and 3.4 of
rufus-scheduler. Sadly, some backwards-incompatible changes were
introduced in this minor versions which must be corrected before
ManageIQ can used the latest released version.

This commit backports Ruby 2.4 stuff to v3.1.10, which we will use in
ManageIQ until we either update rufus-scheduler or find a replacement.
Specifically, the change to use EoTime from the EtOrbi gem does not work
with our stuff as we assume you can use ActiveSupport::Duration and
other such classes in the rufus-scheduler interface.

This situation is compounded as currently tests are indeterminately
failing locally on users of MacOS's.